+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+- Added `rust-wrong-type-user-ptr` to `wrong-type-argument` classification.
 
 ## [0.10.3] - 2019-07-24
 - Made `#[defun]` function signatures display correctly in `help-mode` and `helpful-mode`.